refactor: refactor stores
