+++ src/store/cart/cart.store.ts
@@ -1,13 +1,12 @@
 import { defineStore } from 'pinia';
-import { ref, computed, watch, onBeforeMount } from 'vue';
+import { ref, computed, watch } from 'vue';
 import { IProduct } from '@/services//model/product';
-import { ICartProduct, TProductFunc, TFindFunc, TCurrProd } from './_types';
+import { ICartItem, TProductFunc, TFindFunc, TCurrProd, CartItem } from './_types';
 import { LocalStorageApi } from '@/services/local-storage';
-import { CartProduct } from './_types';
 import { CartDefaultVal, LSKey } from './_constants';
 
 export const useCart = defineStore('cart', () => {
-  const _cart = ref<ICartProduct[]>([]);
+  const _cart = ref<ICartItem[]>([]);
   const _LS = LocalStorageApi.getInstance();
 
   const totalProducts = computed((): number => {
@@ -16,84 +15,80 @@ export const useCart = defineStore('cart', () => {
     }, 0);
   });
 
-  const cart = computed((): ICartProduct[] => [..._cart.value]);
+  const cart = computed((): ICartItem[] => [..._cart.value]);
 
   const clearCart = (): void => {
     _cart.value = [];
   };
 
-  const findProduct: TFindFunc<ICartProduct> = (id) => {
-    return _cart.value.find((product) => {
-      return product.id === id;
+  const findProduct: TFindFunc<ICartItem> = (id) => {
+    return _cart.value.find((item) => {
+      return item.product.id === id;
     });
   };
 
-  const addProduct = (incomeProduct: IProduct, incomeCount = CartDefaultVal.ProductCount): void => {
-    if (!findProduct(incomeProduct.id)) {
-      const cartProduct: ICartProduct = new CartProduct(incomeProduct, incomeCount);
+  const addProduct = (product: IProduct, count = CartDefaultVal.ProductCount): void => {
+    if (!findProduct(product.id)) {
+      const cartProduct: ICartItem = new CartItem(product, count);
       _cart.value.push(cartProduct);
     }
   };
 
-  const dropProduct = (id: string): void => {
+  const dropProduct = (id: IProduct['id']): void => {
     if (findProduct(id)) {
-      _cart.value = _cart.value.filter((prod) => id !== prod.id);
+      _cart.value = _cart.value.filter((item) => id !== item.product.id);
     }
   };
 
-  const incrementCount: TProductFunc = (id: string) => {
+  const incrementCount: TProductFunc = (id) => {
     const currProduct: TCurrProd = findProduct(id);
 
     if (currProduct) currProduct.count += 1;
   };
 
-  const decrementCount: TProductFunc = (id: string) => {
+  const decrementCount: TProductFunc = (id) => {
     const currProduct: TCurrProd = findProduct(id);
 
     if (currProduct) {
       if (currProduct.count < CartDefaultVal.decrementLimit) {
-        _cart.value = _cart.value.filter((prod) => id !== prod.id);
+        _cart.value = _cart.value.filter((item) => id !== item.product.id);
       } else {
         currProduct.count -= 1;
       }
     }
   };
 
-  const updateCount = (val: string, id: string) => {
-    const currProduct: TCurrProd = findProduct(id);
+  const updateCount = (val: string, id: IProduct['id']) => {
+    const currItem: TCurrProd = findProduct(id);
     const valNumber = Number(val);
 
-    if (currProduct) {
+    if (currItem) {
       if (!valNumber) {
         dropProduct(id);
-        return
-      }
-      if (valNumber > currProduct.stock) {
-        currProduct.count = currProduct.stock;
         return;
       }
-      currProduct.count = valNumber;
+      if (valNumber > currItem.product.stock) {
+        currItem.count = currItem.product.stock;
+        return;
+      }
+      currItem.count = valNumber;
     }
   };
 
   watch(
-    () => _cart,
+    _cart,
     (newCart) => {
-      _LS.setProperty(LSKey.cart, newCart.value);
+      _LS.setProperty(LSKey.cart, newCart);
     },
     { deep: true },
   );
 
-  onBeforeMount(() => {
+  (function initStore() {
     const cartLS: unknown = _LS.getProperty(LSKey.cart);
-    if (cartLS instanceof Array<ICartProduct>) {
-      const newCart: ICartProduct[] = cartLS;
-      _cart.value = [];
-      newCart.forEach((product) => {
-        addProduct(product, product.count);
-      });
+    if (cartLS instanceof Array<ICartItem>) {
+      cartLS.forEach((item) => addProduct(item.product, item.count));
     }
-  });
+  })();
 
   return {
     cart,
